refactor(katasterui): flatter file structures
