+++ src/pages/KatasterKI/components/QuestionTypes/Info.js
