+++ src/pages/KatasterKI/components/QuestionTypes/SingleChoice.js
