+++ src/pages/KatasterKI/components/QuestionTypes/Sliders.js
